+++ src/entities/api-key/actions.ts
@@ -1,0 +1,5 @@
+'use server';
+
+export { saveApiKeyAction } from './actions/saveApiKeyAction';
+export { deleteApiKeyAction } from './actions/deleteApiKeyAction';
+export { getRegisteredProvidersAction } from './actions/getRegisteredProvidersAction';
